changed slice behavior for multiple hotkeys entered
There's a failsafe for the uncommon case of entering more than one character into a hotkey field (which normally only allows one because it re-selects the entire field every time). The script will now select the last character entered instead of first, which is more predictable and consistent with normal behavior.