added option to allow position conflict during drop
If the shift key is held when a button is dropped, the editor will now place the button there without swapping positions with the button already in that position. This creates a button conflict, which may be desirable in certain situations (the game and the editor both overflow one of the buttons into an adjacent position, depending on the order of the commands in the game data).